Feature ab126 api data user specific

+++ nba-api-dotnet/Controllers/TeamsController.cs
@@ -30,6 +30,25 @@ public class TeamController : ControllerBase
     {
         try
         {
+            var sub = HttpContext?.User.Claims.FirstOrDefault(c => c.Type == System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+
+            User? isUser = await _context.tbl_Users
+            .Where(t => t.UserIdentifier == sub)
+            .FirstOrDefaultAsync();
+
+            if(isUser == null){
+                User UserToAdd = new User();
+                UserToAdd.UserIdentifier=sub;
+                await _context.tbl_Users.AddAsync(UserToAdd);
+                await _context.SaveChangesAsync();
+
+                isUser = await _context.tbl_Users
+                .Where(t => t.UserIdentifier == sub)
+                .FirstOrDefaultAsync();
+            }
+
+            team.UserId=isUser.UserID;
+
             if (team.TeamName is string)
             {
                 team.TeamName = team.TeamName.TrimStart();
@@ -44,6 +63,7 @@ public class TeamController : ControllerBase
             //Check that team doesn't already exist in the database
             Team? isTeam = await _context.tbl_Teams
             .Where(t => t.TeamName == team.TeamName)
+            .Where(t => t.UserId == isUser.UserID )
             .FirstOrDefaultAsync();
 
 
@@ -70,27 +90,27 @@ public class TeamController : ControllerBase
     }
 
 
-    /// <summary>
-    /// Gets a List of all the teams in the database
-    /// </summary>
-    /// <returns>A list of Teams</returns>
-    [HttpGet]
-    [Authorize]
-    [Route("get-all")]
-    public async Task<ActionResult<List<Team>>> getAllTeams()
-    {
-        try
-        {
-            List<Team> teams = await _context.tbl_Teams.ToListAsync();
+    // <summary>
+    // Gets a List of all the teams in the database
+    // </summary>
+    // <returns>A list of Teams</returns>
+    // [HttpGet]
+    // [Authorize]
+    // [Route("get-all")]
+    // public async Task<ActionResult<List<Team>>> getAllTeams()
+    // {
+    //     try
+    //     {
+    //         List<Team> teams = await _context.tbl_Teams.ToListAsync();
 
-            return Ok(teams);
-        }
-        catch (Exception ex)
-        {
-            return StatusCode(500, ex.ToString());
-        }
+    //         return Ok(teams);
+    //     }
+    //     catch (Exception ex)
+    //     {
+    //         return StatusCode(500, ex.ToString());
+    //     }
 
-    }
+    // }
 
     /// <summary>
     /// Get Players on a specific team
@@ -147,6 +167,17 @@ public class TeamController : ControllerBase
     {
         try
         {
+            var sub = HttpContext?.User.Claims.FirstOrDefault(c => c.Type == System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+
+            User? isUser = await _context.tbl_Users
+            .Where(t => t.UserIdentifier == sub)
+            .FirstOrDefaultAsync();
+            if (isUser == null ) {
+                var response = new Response<Team?>(null, false, "This User Does Not Exist");
+
+                return StatusCode(409, response);
+            }
+            
             var teamID = Convert.ToInt32(RouteData.Values["teamID"]);
 
             //check to see if team exists
@@ -157,6 +188,13 @@ public class TeamController : ControllerBase
             if (isTeam == null)
             {
                 var response = new Response<Team?>(null, false, "This Team does not exist");
+
+                return StatusCode(409, response);
+            }
+
+            if (isTeam.UserId != isUser.UserID)
+            {
+                var response = new Response<Team?>(null, false, "This Team is not assigned to this user");
 
                 return StatusCode(409, response);
             }
@@ -222,8 +260,19 @@ public class TeamController : ControllerBase
     {
         try
         {
-            var teamID = Convert.ToInt32(RouteData.Values["teamID"]!);
             var response = new Response<List<int?>>();
+            var sub = HttpContext?.User.Claims.FirstOrDefault(c => c.Type == System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+
+            User? isUser = await _context.tbl_Users
+            .Where(t => t.UserIdentifier == sub)
+            .FirstOrDefaultAsync();
+            if (isUser == null ) {
+                response = new Response<List<int?>>(new List<int?>(), false, "This User Does Not Exist");
+
+                return StatusCode(409, response);
+            }
+            var teamID = Convert.ToInt32(RouteData.Values["teamID"]!);
+            
 
             //check to see if team exists
             var isTeam = await _context.tbl_Teams
@@ -233,6 +282,13 @@ public class TeamController : ControllerBase
             if (isTeam == null)
             {
                 response = new Response<List<int?>>(new List<int?>(), false, "This Team does not exist");
+
+                return StatusCode(409, response);
+            }
+
+            if (isTeam.UserId != isUser.UserID)
+            {
+                response = new Response<List<int?>>(new List<int?>(), false, "This Team is not assigned to this user");
 
                 return StatusCode(409, response);
             }
@@ -293,8 +349,20 @@ public class TeamController : ControllerBase
     {
         try
         {
+           
             var teamID = Convert.ToInt32(RouteData.Values["teamID"]!);
             var response = new Response<int>();
+            
+            var sub = HttpContext?.User.Claims.FirstOrDefault(c => c.Type == System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+
+            User? isUser = await _context.tbl_Users
+            .Where(t => t.UserIdentifier == sub)
+            .FirstOrDefaultAsync();
+            if (isUser == null ) {
+                response = new Response<int>(0, false, "This User Does Not Exist");
+
+                return StatusCode(409, response);
+            }
 
             //check to see if team exists
             var isTeam = await _context.tbl_Teams
@@ -308,7 +376,12 @@ public class TeamController : ControllerBase
                 return StatusCode(409, response);
             }
 
+             if (isTeam.UserId != isUser.UserID)
+            {
+                response = new Response<int>(0, false, "This Team is not assigned to this user");
 
+                return StatusCode(409, response);
+            }
 
             // if team exists and all players to delete are on team, do so
             _context.tbl_PlayerSelection.RemoveRange(
@@ -345,7 +418,9 @@ public class TeamController : ControllerBase
         
         try
         {
-            var teams = await _context.view_WinChance.ToListAsync();
+            var sub = HttpContext?.User.Claims.FirstOrDefault(c => c.Type == System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+            
+            var teams = await _context.view_WinChance.Where(t => t.UserIdentifier == sub).ToListAsync();
 
             var response = new Response<List<WinChanceView>>(teams, true, "Team Successfully returned");
